- seperate cmds - check_equilibrium - graphstatics comments spelling mistake
